services/mqtt-nats-bridge/requirements.txt
# Maestra MQTT-NATS Bridge - Python Dependencies

# MQTT Protocol
paho-mqtt==2.0.0

# NATS Protocol
nats-py==2.6.0

# Utilities
python-dotenv==1.0.0
